fix package names
