+++ trunk/source/Library-TraceCheckerUtils/META-INF/MANIFEST.MF
@@ -7,11 +7,11 @@ Bundle-RequiredExecutionEnvironment: JavaSE-11
 Export-Package: 
  de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils,
  de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.cfg2automaton,
+ de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.partialorder,
+ de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.partialorder.independence,
+ de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.partialorder.petrinetlbe,
  de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.predicates,
- de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck,
- de.uni_freiburg.informatik.ultimate.lib.tracecheckutils.partialorder,
- de.uni_freiburg.informatik.ultimate.lib.tracecheckutils.partialorder.independence,
- de.uni_freiburg.informatik.ultimate.lib.tracecheckutils.partialorder.petrinetlbe
+ de.uni_freiburg.informatik.ultimate.lib.tracecheckerutils.singletracecheck
 Require-Bundle: de.uni_freiburg.informatik.ultimate.lib.core,
  de.uni_freiburg.informatik.ultimate.lib.smtlib,
  de.uni_freiburg.informatik.ultimate.lib.automata,
